Implemented paragraph based nodes and links
Changed the DB generation to stream the BGE-M3 embeddings and match them with the pages from the Wikipedia XML dump. Instead of pages, we set paragraphs as the basic node type and link them to each other. References are extracted for each paragraph.
The basic structure of the graph is that the links between pages point to the first paragraph (which is usually a summary). The first paragraph also links to all other paragraphs on the page. Other paragraphs only link to the next one.